Safely update _isLoading state
Because `_loadPage` is called within lifecycle methods (didChangeDependencies, didUpdateWidget) and it is an async method, there is a possibility that it will execute `_loadPage` and then unmount before the `setState` methods after the `await` calls have been executed. When `setState` is called on an unmounted widget, an error is thrown and the app crashes.

This PR checks if the component is still mounted before it calls the `setState` method after the `await` calls.